Generalized structure for adding new conversations + epistemic queerness
