+++ components/Queerness.tsx
@@ -1,0 +1,10 @@
+import QueernessData from '../data/queerness.json';
+import Example from './Example';
+
+const Queerness = () => {
+  return (
+    <Example data={QueernessData} />
+  );
+};
+
+export default Queerness;
